+++ src/app/utils/items-select-browser/filesystem-model.js
@@ -162,6 +162,17 @@ export default BaseModel.extend(I18n, {
   /**
    * @override
    */
+  willDestroy() {
+    try {
+      this.browserModel?.destroy?.();
+    } finally {
+      this._super(...arguments);
+    }
+  },
+
+  /**
+   * @override
+   */
   async fetchChildren(dirId, startIndex, size, offset) {
     const fileManager = this.get('fileManager');
     return fileManager